Managers

+++ C_Plus_Plus_Game/HealthUi.h
@@ -4,7 +4,6 @@
 #include "GameObject.h"
 #include "Util.h"
 #include "GameState.h"
-
 class HealthUI :public GameObject
 {
 protected:
@@ -47,7 +46,7 @@ public:
 		m_height = height;
 	}
 
-	void updateUIOnDamage(const float& initialHealth, const  float& currentHealth)
+	void updateUIOnDamage(const int& initialHealth, const  int& currentHealth)
 	{
 		m_startTime = *m_state->getPausableClock();
 		//reduces fill size, so that the red background appears
